texdreamer/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
chumpy==0.70
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.58.1
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.6.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.4
imageio==2.37.0
importlib-metadata==8.7.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.8
markupsafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook-shim==0.2.4
numpy==1.26.4
open3d-python==0.3.0.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.1.1
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pyglet==2.1.6
pygments==2.19.1
pyopengl==3.1.0
pyparsing==3.2.3
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch3d==0.7.5
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
safetensors==0.5.3
scipy==1.15.3
send2trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==3.1.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trimesh==4.6.10
triton==2.1.0
types-python-dateutil==2.9.0.20250516
typing-extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.13
virtualenv==16.7.9
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xformers==0.0.22.post7+cu118
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zipp==3.22.0
